Consolidate CSS border styles for subscriber accordion components
