+++ src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberInfo/Summary.tsx
@@ -1,4 +1,4 @@
-import type { SubscriberSummaryProps } from 'src/components/admin/Settings/PrefixAlerts/types';
+import type { SubscriberAccordionProps } from 'src/components/admin/Settings/PrefixAlerts/types';
 import type { ChipDisplay } from 'src/components/shared/ChipList/types';
 import type { AlertTypeInfo } from 'src/gql-types/graphql';
 
@@ -16,14 +16,9 @@ import { NavArrowDown } from 'iconoir-react';
 
 import ChipList from 'src/components/shared/ChipList';
 import { useGetAlertTypes } from 'src/context/AlertType';
-import { defaultOutline, defaultOutlineColor_hovered } from 'src/context/Theme';
 import { sortByAlertType } from 'src/utils/misc-utils';
 
-const Summary = ({
-    subscription,
-    expanded,
-    setHovered,
-}: SubscriberSummaryProps) => {
+const Summary = ({ expanded, subscription }: SubscriberAccordionProps) => {
     const theme = useTheme();
 
     const [alertTypeResponse] = useGetAlertTypes();
@@ -68,22 +63,9 @@ const Summary = ({
             expandIcon={
                 <NavArrowDown style={{ color: theme.palette.text.primary }} />
             }
-            onMouseOver={() => setHovered(true)}
-            onMouseOut={() => setHovered(false)}
             sx={{
-                'backgroundColor':
-                    theme.palette.mode === 'dark' ? 'transparent' : 'white',
-                'border': defaultOutline[theme.palette.mode],
-                'borderBottom': expanded ? 'none' : undefined,
-                'borderRadius': 3,
-                'borderBottomLeftRadius': expanded ? 'unset' : undefined,
-                'borderBottomRightRadius': expanded ? 'unset' : undefined,
                 'minHeight': 40,
                 'px': 1,
-                '&:hover': {
-                    borderColor:
-                        defaultOutlineColor_hovered[theme.palette.mode],
-                },
                 '&:hover::after': {
                     backgroundColor: 'transparent',
                 },
