+++ src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberInfo/index.tsx
@@ -2,28 +2,37 @@ import type { SubscriberInfoProps } from 'src/components/admin/Settings/PrefixAl
 
 import { useState } from 'react';
 
-import { Accordion } from '@mui/material';
+import { Accordion, useTheme } from '@mui/material';
 
 import Details from 'src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberInfo/Details';
 import Summary from 'src/components/admin/Settings/PrefixAlerts/Dialog/SubscriberInfo/Summary';
+import { defaultOutline, defaultOutlineColor_hovered } from 'src/context/Theme';
 
 const SubscriberInfo = ({ subscription }: SubscriberInfoProps) => {
+    const theme = useTheme();
+
     const [expanded, setExpanded] = useState(false);
-    const [hovered, setHovered] = useState(false);
 
     return (
-        <Accordion expanded={expanded} onChange={() => setExpanded(!expanded)}>
-            <Summary
-                subscription={subscription}
-                expanded={expanded}
-                setHovered={setHovered}
-            />
+        <Accordion
+            expanded={expanded}
+            onChange={() => setExpanded(!expanded)}
+            sx={{
+                'backgroundColor':
+                    theme.palette.mode === 'dark' ? 'transparent' : 'white',
+                'border': defaultOutline[theme.palette.mode],
+                '&:hover': {
+                    borderColor:
+                        defaultOutlineColor_hovered[theme.palette.mode],
+                },
+                [`&:first-of-type`]: {
+                    borderRadius: '6px',
+                },
+            }}
+        >
+            <Summary subscription={subscription} expanded={expanded} />
 
-            <Details
-                subscription={subscription}
-                expanded={expanded}
-                hovered={hovered}
-            />
+            <Details subscription={subscription} />
         </Accordion>
     );
 };
